Change example to use a catch all
This example builds on the previous commit, instead of rendering one
specific page all routes that aren't handled explicitly in our
application will fallback to builder.

You need to have a model named "page" - when creating a new page you
should specify that it targets a specific url. Visit this url in the app
to see your content.